Séparation des tests en deux fichiers, et ajout du fichier UserListViewModelTests
